refactor(mainnet): add pallet_scheduler to utility
